requirements.txt
-i https://pypi.org/simple
attrs==22.2.0 ; python_version >= '3.6'
beautifulsoup4==4.11.1
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_version >= '3.6'
colorama==0.4.6 ; sys_platform == 'win32'
exceptiongroup==1.1.0 ; python_version < '3.11'
idna==3.4 ; python_version >= '3.5'
iniconfig==1.1.1
packaging==22.0 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
psycopg2==2.9.5
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
requests-mock==1.10.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.2.post1 ; python_version >= '3.6'
tomli==2.0.1 ; python_version < '3.11'
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
